% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{create_fish_pop}
\alias{create_fish_pop}
\title{Create fish population}
\usage{
create_fish_pop(
  nfish = 5000,
  init_fl = list(min = 28, max = 80, mode = 38),
  fl_distribution = "triangle",
  psizeclass = c(0.3, 0.5, 0.2),
  ...
)
}
\arguments{
\item{nfish}{number of fish in the population}

\item{init_fl}{Named list ('min', 'max', 'mode') of initial fork lengths when
distribution of fish for fl_distribution = 'triangle' (default)}

\item{fl_distribution}{'triangle' uses rtriangle() distribution.
'discrete' creates three discrete size classes ('small', 'medium', 'large')
according to psizeclass}

\item{psizeclass}{proportion in each size class when fl_distribution = 'discrete'}

\item{...}{}
}
\value{

}
\description{
Function to create fish population with specified abundance and
fork length distribution.
}
