% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{grow_sum_fish}
\alias{grow_sum_fish}
\title{Title}
\usage{
grow_sum_fish(filled_reach_list, ...)
}
\arguments{
\item{filled_reach_list}{list generated by fill_reach function}

\item{...}{}
}
\value{

}
\description{
Title
}
