% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{make_reach}
\alias{make_reach}
\title{Make reach}
\usage{
make_reach(reach_in, ag_zoop, river_zoop, ag_temp, river_temp, vdays)
}
\arguments{
\item{reach_in}{reach}

\item{ag_zoop}{prey density (individuals per liter) in ag field water}

\item{river_zoop}{prey density in river}

\item{ag_temp}{water temperature of ag field water}

\item{river_temp}{water temperature of river}

\item{vdays}{day(s) on which mass will be calculated.}
}
\value{

}
\description{
Assign temperature and prey density to cells in reach
}
