% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{newmass_fxn}
\alias{newmass_fxn}
\title{Calculate new mass}
\usage{
newmass_fxn(df, old_mass, day)
}
\arguments{
\item{df}{a dataframe with the columns "temperature" and "prey_density"}

\item{old_mass}{initial mass of fish in grams}

\item{day}{day on which mass should be calculated}
}
\value{
Vector
}
\description{
Calculate the new mass of a fish from its old mass after a specficed number of days, prey density, and temperature
}
