% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{R_fxn}
\alias{R_fxn}
\title{Ration function}
\usage{
R_fxn(
  foraging_time = FORAGING_TIME,
  consumption_rate,
  dry_mass = DRY_MASS,
  Temperature,
  M0,
  ...
)
}
\arguments{
\item{foraging_time}{time (in minutes) a fish forages. Default = 360}

\item{consumption_rate}{calculated with typeII function}

\item{dry_mass}{dry mass (in ___) of prey}

\item{Temperature}{water temperature (in C)}

\item{M0}{initial fish mass (in grams)}

\item{...}{}
}
\value{

}
\description{
Calculate the ration available to a fish given a consumption rate,
foraging time, and prey dry mass. Available ration is relative to the
temperature- and mass-dependent ration that maximizes growth. When
rations >1 food is not limiting to growth; when ration <1 food is limiting.
}
