% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{typeII}
\alias{typeII}
\title{Type II functional feeding response}
\usage{
typeII(
  C_max = max_consumption_rate,
  prey_density,
  P_Chalf = prey_density_at_half_max_consumption,
  ...
)
}
\arguments{
\item{C_max}{max consumption rate (Beta 0)}

\item{prey_density}{Prey density (individuals per liter)}

\item{P_Chalf}{prey density at which consumption reaches half of its maximum (Beta 1)}

\item{...}{}
}
\value{

}
\description{
Calculate the consumption rate using the Type II functional response from
Haskell et al 2017 using sub-yearling Chinook and Daphnia.
}
