% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{W2L_fxn}
\alias{W2L_fxn}
\title{Convert mass to fork length}
\usage{
W2L_fxn(mass, a = 0.005, b = 3.44)
}
\arguments{
\item{mass}{mass to convert}

\item{a}{coefficient}

\item{b}{coefficient}
}
\value{

}
\description{
Convert mass to fork length
}
