% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fuctions.R
\name{newmass_fxn}
\alias{newmass_fxn}
\title{Calculate new mass}
\usage{
newmass_fxn(prey_density, temperature, old_mass, day)
}
\arguments{
\item{prey_density}{prey_density in individuals per liter}

\item{temperature}{water temperature in degrees C}

\item{old_mass}{initial mass of fish in grams}

\item{day}{day on which mass should be calculated}
}
\value{
Vector
}
\description{
Calculate the new mass of a fish from its old mass after a specficed number of days, prey density, and temperature
}
